pyproject.toml
[tool.poetry]
name = "oz-03-collabo-004-be"
version = "0.1.0"
description = ""
authors = ["dayeonkimm <ekdyd516@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
psycopg2-binary = "^2.9"  # PostgreSQL을 사용하기 위한 패키지
python-dotenv = "^1.0"  # .env 파일을 로드하기 위한 패키지
djangorestframework = "^3.14" 
djangorestframework-simplejwt = "^5.3.1"
pytz = "*"
boto3 = "*"
django-storages = "*"
coverage = "^7.6.1"
django-cors-headers = "^3.14.0"
pillow = "^10.4.0"
openai = "^1.42.0"
locust = "^2.31.5"




[tool.poetry.group.dev.dependencies]
coverage = "^7.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
